+++ core/src/main/kotlin/com/gitlab/kordlib/core/live/channel/LiveGuildChannel.kt
@@ -1,0 +1,32 @@
+package com.gitlab.kordlib.core.live.channel
+
+import com.gitlab.kordlib.common.annotation.KordPreview
+import com.gitlab.kordlib.core.entity.Entity
+import com.gitlab.kordlib.core.entity.channel.GuildChannel
+import com.gitlab.kordlib.core.entity.channel.GuildMessageChannel
+import com.gitlab.kordlib.core.event.Event
+import com.gitlab.kordlib.core.event.channel.ChannelCreateEvent
+import com.gitlab.kordlib.core.event.channel.ChannelDeleteEvent
+import com.gitlab.kordlib.core.event.channel.ChannelUpdateEvent
+import com.gitlab.kordlib.core.event.guild.GuildDeleteEvent
+
+@KordPreview
+fun GuildChannel.live() = LiveGuildChannel(this)
+
+@KordPreview
+class LiveGuildChannel(channel: GuildChannel) : LiveChannel(), Entity by channel {
+
+    override var channel: GuildChannel = channel
+        private set
+
+    override fun update(event: Event) = when (event) {
+        is ChannelCreateEvent -> channel = event.channel as GuildMessageChannel
+        is ChannelUpdateEvent -> channel = event.channel as GuildMessageChannel
+        is ChannelDeleteEvent -> shutDown()
+
+        is GuildDeleteEvent -> shutDown()
+
+        else -> Unit
+    }
+
+}